+++ src/atoms/layout/page/index.module.scss
@@ -6,7 +6,7 @@
 
   margin: auto;
   min-height: 100vh;
-  // min-width: 320px;
+  /* min-width: 320px;*/
   width: 100%;
 
   display: flex;
@@ -24,7 +24,7 @@
     max-width: var(--max-width-internal);
   }
 
-  // container: page / inline-size;
+  /* container: page / inline-size;*/
 
   &.feed {
     .content {
