+++ src/atoms/tab/Tabs.tsx
@@ -2,29 +2,13 @@ import { useMemo } from 'react'
 import styles from '@style'
 import { Line } from '@atoms/line'
 import Button from '@atoms/button/Button'
-export interface TabOptions {
-  title: string
-  to?: string
-  disabled?: boolean
-  /**mod overrides */
-  restricted?: boolean
-  /**only relevant when user is the owner*/
-  private?: boolean
-}
-
-interface TabsProps {
-  tabs: TabOptions[]
-  className?: string
-  /**A method to filter out the tabs */
-  filter?: (tabs: TabOptions) => TabOptions | null
-}
 
 /**
  * Core Tabs
  */
 export const Tabs = ({
   tabs,
-  className,
+  className = '',
   filter /*onClickprops = {}*/,
 }: TabsProps) => {
   const filtered_tabs = useMemo(
@@ -33,7 +17,7 @@ export const Tabs = ({
   )
 
   return (
-    <div className={`${styles.container} ${className ? className : ''}`}>
+    <div className={`${styles.container} ${className}`}>
       <div className={styles.tabs}>
         {filtered_tabs.map((tab, index) => {
           if (tab) {
@@ -41,7 +25,7 @@ export const Tabs = ({
               <Button
                 alt={`tab_${tab.title}`}
                 preventScrollReset
-                key={tab.title}
+                key={index}
                 className={styles.tab}
                 activeClass={styles.active}
                 disabled={tab.disabled}
