+++ src/atoms/toggles/Toggle.tsx
@@ -4,17 +4,6 @@ import type React from 'react'
 import { useCallback } from 'react'
 import { useControlled } from '@hooks/use-controlled'
 
-interface ToggleProps {
-  label?: string
-  initial?: boolean
-  toggled?: boolean
-  style?: React.CSSProperties
-  alt?: string
-  box?: boolean
-  onToggle?: (v: boolean) => void
-  className?: string
-}
-
 export const Toggle = ({
   label,
   initial,
