+++ src/data/api.ts
@@ -92,22 +92,6 @@ const GetUserClaims = async (walletAddr: string) => {
   })
 }
 
-interface TzpMetadata {
-  alias?: string
-  tzprofile?: string
-  twitter?: string
-  discord?: string
-  github?: string
-  dns?: string
-}
-interface TzktMetadata {
-  twitter?: string
-  alias?: string
-}
-
-interface TzktData {
-  data?: TzktMetadata
-}
 /**
  * Get User Metadata
  */
@@ -117,7 +101,7 @@ export const GetUserMetadata = async (walletAddr: string) => {
   const tzpData: TzpMetadata = {}
   try {
     const claims = await GetUserClaims(walletAddr)
-    if (claims.data.data.tzprofiles_by_pk !== null)
+    if (claims.data.data.tzprofiles_by_pk !== null) {
       for (const claim of claims.data.data.tzprofiles_by_pk.valid_claims) {
         const claimJSON = JSON.parse(claim[1])
         if (claimJSON.type.includes('TwitterVerification')) {
@@ -125,8 +109,12 @@ export const GetUserMetadata = async (walletAddr: string) => {
             tzpData.twitter = claimJSON.evidence.handle
           }
         } else if (claimJSON.type.includes('BasicProfile')) {
-          if (claimJSON.credentialSubject.alias !== '' && !tzktData.data?.alias)
+          if (
+            claimJSON.credentialSubject.alias !== '' &&
+            !tzktData.data?.alias
+          ) {
             tzpData.alias = claimJSON.credentialSubject.alias
+          }
           tzpData.tzprofile = walletAddr
         } else if (claimJSON.type.includes('DiscordVerification')) {
           if (!tzktData.data) {
@@ -143,6 +131,7 @@ export const GetUserMetadata = async (walletAddr: string) => {
           tzpData.dns = claimJSON.credentialSubject.sameAs.slice(4)
         }
       }
+    }
   } catch (e: any) {
     console.error(e, e.stack)
   }
