+++ src/data/ipfs.ts
@@ -1,4 +1,4 @@
-import { MIMETYPE, IPFS_DEFAULT_THUMBNAIL_URI } from '@constants'
+import { IPFS_DEFAULT_THUMBNAIL_URI, MIMETYPE } from '@constants'
 
 import mime from 'mime-types'
 import axios from 'axios'
@@ -423,76 +423,4 @@ async function buildMetadataFile({
   const cleaned_metas = pickBy(metadata, (v) => v !== undefined)
 
   return JSON.stringify(cleaned_metas)
-}
-
-// TODO: Move types to each roots?
-
-interface PrepareProps {
-  name: string
-  description: string
-  tags: string
-  address: string
-  file: FileForm
-  cover: FileForm
-  thumbnail: FileForm
-  rights: string
-  rightUri?: string
-  language?: string
-  accessibility: string
-  contentRating: string
-  formats: any
-}
-
-interface PrepareDirectoryOptions {
-  name?: string
-  description?: string
-  tags?: string
-  address: string
-  files: FileMint[]
-  cover: FileForm
-  thumbnail: FileForm
-  formats: MintFormat[]
-  generateDisplayUri: string
-  rights?: string
-  rightUri?: string
-  language?: string
-  accessibility?: string
-  contentRating?: string
-}
-interface BuildMetadataOptions {
-  name?: string
-  description?: string
-  tags?: string
-  uri: string
-  address: string
-  displayUri: string
-  thumbnailUri: string
-  formats: MintFormat[]
-  rights?: string
-  rightUri?: string
-  language?: string
-  accessibility?: string
-  contentRating?: string
-}
-interface TeiaMetadata {
-  name: string
-  description: string
-  tags: string[]
-  symbol: string
-  artifactUri: string
-  displayUri: string
-  thumbnailUri: string
-  creators: string[]
-  formats: MintFormat[]
-  decimals: number
-  isBooleanAmount: boolean
-  shouldPreferSymbol: boolean
-  rights: string
-  date: string
-  mintingTool: string
-  //optional
-  accessibility?: string
-  contentRating?: string
-  rightUri?: string
-  language?: string
 }
