+++ src/utils/with-icon.tsx
@@ -11,7 +11,15 @@ interface WithIconProps {
   viewBox?: number
 }
 
-// SVG Icon Cache system using refs
+/** SVG Icon Cache system using refs
+ * Resolves the first call to the actuall SVG and store it in a
+ * context cache. We then `useref` that svg on subsequent calls.
+ *
+ * Less optimal to use than importing SVG using the bundler (as a react component)
+ * for single use cases.
+ *
+ * Heavily based on https://paco.me/writing/svg-caching-with-use
+ */
 const withIcon = (icon: string, defaultViewBox?: number) => {
   const Icon = (props: WithIconProps) => {
     const { size = 30, viewBox = defaultViewBox } = props
